# planner: don't calc the heavy expression used in ORDER BY stmt twice