--- a/pkg/planner/core/casetest/vectorsearch/BUILD.bazel
+++ b/pkg/planner/core/casetest/vectorsearch/BUILD.bazel
@@ -9,7 +9,7 @@ go_test(
     ],
     data = glob(["testdata/**"]),
     flaky = True,
-    shard_count = 5,
+    shard_count = 7,
     deps = [
         "//pkg/config",
         "//pkg/domain",