… pkg/planner/core/casetest/vectorsearch/testdata/ann_index_suite_in.json
… pkg/planner/core/casetest/vectorsearch/testdata/ann_index_suite_out.json