<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #54245 , close #56318 

Problem Summary:
cherry-pick : https://github.com/tidbcloud/tidb-cse/pull/1426

In origin plan, root will calculate distance when using vector search although the result has been calculated in store node.  For example:
```sql
mysql> explain SELECT id FROM t ORDER BY vec_l2_distance(val, '[1,3,6]') LIMIT 2;
+----------------------------------+---------+-----------+---------------+-----------------------------------------------------------------------+
| id                               | estRows | task      | access object | operator info                                                         |
+----------------------------------+---------+-----------+---------------+-----------------------------------------------------------------------+
| Projection_7                     | 2.00    | root      |               | test.t.id                                                             |
| └─Projection_15                  | 2.00    | root      |               | test.t.id, test.t.val                                                 |
|   └─TopN_8                       | 2.00    | root      |               | Column#4, offset:0, count:2                                           |
|     └─Projection_16              | 2.00    | root      |               | test.t.id, test.t.val, vec_l2_distance(test.t.val, [1,3,6])->Column#4 |
|       └─TableReader_14           | 2.00    | root      |               | data:TopN_13                                                          |
|         └─TopN_13                | 2.00    | cop[tikv] |               | vec_l2_distance(test.t.val, [1,3,6]), offset:0, count:2               |
|           └─TableFullScan_12     | 3.00    | cop[tikv] | table:t       | keep order:false, stats:pseudo                                        |
+----------------------------------+---------+-----------+---------------+-----------------------------------------------------------------------+
7 rows in set (0.01 sec)

mysql> explain SELECT id FROM t2 ORDER BY vec_cosine_distance(val, '[1,3,6]') LIMIT 2;
+----------------------------------------+---------+--------------+------------------------------------+------------------------------------------------------------------------------+
| id                                     | estRows | task         | access object                      | operator info                                                                |
+----------------------------------------+---------+--------------+------------------------------------+------------------------------------------------------------------------------+
| Projection_7                           | 2.00    | root         |                                    | test.t2.id                                                                   |
| └─Projection_29                        | 2.00    | root         |                                    | test.t2.id, test.t2.val                                                      |
|   └─TopN_11                            | 2.00    | root         |                                    | Column#5, offset:0, count:2                                                  |
|     └─Projection_30                    | 2.00    | root         |                                    | test.t2.id, test.t2.val, vec_cosine_distance(test.t2.val, [1,3,6])->Column#5 |
|       └─TableReader_26                 | 2.00    | root         |                                    | MppVersion: 1, data:ExchangeSender_25                                        |
|         └─ExchangeSender_25            | 2.00    | mpp[tiflash] |                                    | ExchangeType: PassThrough                                                    |
|           └─Projection_27              | 2.00    | mpp[tiflash] |                                    | test.t2.id, test.t2.val                                                      |
|             └─TopN_24                  | 2.00    | mpp[tiflash] |                                    | Column#4, offset:0, count:2                                                  |
|               └─Projection_28          | 2.00    | mpp[tiflash] |                                    | test.t2.id, test.t2.val, vec_cosine_distance(test.t2.val, [1,3,6])->Column#4 |
|                 └─TableFullScan_23     | 2.00    | mpp[tiflash] | table:t2, index:idx_embedding(val) | keep order:false, stats:pseudo, annIndex:COSINE(val..[1,3,6], limit:2)       |
+----------------------------------------+---------+--------------+------------------------------------+------------------------------------------------------------------------------+
10 rows in set (0.00 sec)
```

After this pr, it can be optimized by reuse the distance column and avoid exchange vector column.
```sql
mysql> explain select id from t1 order by vec_cosine_distance(vec, '[1,1,1]') limit 10;
+----------------------------------+---------+--------------+------------------------------------+------------------------------------------------------------------------------+
| id                               | estRows | task         | access object                      | operator info                                                                |
+----------------------------------+---------+--------------+------------------------------------+------------------------------------------------------------------------------+
| Projection_7                     | 10.00   | root         |                                    | test.t1.id                                                                   |
| └─TopN_11                        | 10.00   | root         |                                    | Column#9, offset:0, count:10                                                 |
|   └─TableReader_28               | 10.00   | root         |                                    | MppVersion: 1, data:ExchangeSender_27                                        |
|     └─ExchangeSender_27          | 10.00   | mpp[tiflash] |                                    | ExchangeType: PassThrough                                                    |
|       └─TopN_26                  | 10.00   | mpp[tiflash] |                                    | Column#9, offset:0, count:10                                                 |
|         └─Projection_25          | 10.00   | mpp[tiflash] |                                    | test.t1.id, test.t1.vec, vec_cosine_distance(test.t1.vec, [1,1,1])->Column#9 |
|           └─TableFullScan_24     | 10.00   | mpp[tiflash] | table:t1, index:idx_embedding(vec) | keep order:false, stats:pseudo, annIndex:COSINE(vec..[1,1,1], limit:10)      |
+----------------------------------+---------+--------------+------------------------------------+------------------------------------------------------------------------------+
7 rows in set (0.00 sec)

mysql> explain SELECT id FROM t ORDER BY vec_l2_distance(val, '[1,3,6]') LIMIT 2;
+--------------------------------+----------+-----------+---------------+-----------------------------------------------------------+
| id                             | estRows  | task      | access object | operator info                                             |
+--------------------------------+----------+-----------+---------------+-----------------------------------------------------------+
| Projection_8                   | 2.00     | root      |               | test.t.id                                                 |
| └─TopN_12                      | 2.00     | root      |               | Column#4, offset:0, count:2                               |
|   └─TableReader_25             | 2.00     | root      |               | data:TopN_24                                              |
|     └─TopN_24                  | 2.00     | cop[tikv] |               | Column#4, offset:0, count:2                               |
|       └─Projection_23          | 10000.00 | cop[tikv] |               | test.t.id, vec_l2_distance(test.t.val, [1,3,6])->Column#4 |
|         └─TableFullScan_20     | 10000.00 | cop[tikv] | table:t       | keep order:false, stats:pseudo                            |
+--------------------------------+----------+-----------+---------------+-----------------------------------------------------------+
6 rows in set (0.00 sec)
```

### What changed and how does it work?
We add `getPushedDownTopN4VectorSearch` to get partial TopN and set a children plan `physicalProjection` to resolve partial TopN distance column. At the same time, apply the column in root plan.

In 768 dimension and 10000 vector data, we test the sql execute time -- `SELECT id FROM table_name ORDER BY Vec_Cosine_Distance(embedding, search_vector) limit 10;`, **about 20+% performance improvement. ⬆️**
```
before: execute time 15.81586ms
after: execute time 12.27633ms
```
### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```